feat/order: orderDetail 추가 및 cart 연동

+++ src/main/java/org/example/eatopia/domain/product/dto/request/ProductUpdateRequest.java
@@ -9,7 +9,7 @@ public record ProductUpdateRequest(
         String description,
         String thumbnailUrl,
         BigDecimal price,
-        Long stock,
+        Integer stock,
         ProductStatus status,
         Long categoryId
 ) {
